requirements.txt
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudscraper==1.2.56
fastapi==0.63.0
flake8==3.8.4
h11==0.12.0
idna==2.10
mccabe==0.6.1
pip==20.3.3
pip-autoremove==0.9.1
pycodestyle==2.6.0
pydantic==1.7.3
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
requests-toolbelt==0.9.1
setuptools==52.0.0.post20210125
six==1.15.0
soupsieve==2.1
starlette==0.13.6
toml==0.10.2
urllib3==1.26.2
uvicorn==0.13.3
wheel==0.36.2
